requirements.txt
配置
PyYAML==3.12
SQLAlchemy==1.1.13
PyMySQL==0.7.11
flower==0.9.1
pymongo==3.5.0

